+++ R/powell_wiley.R
@@ -36,7 +36,7 @@
 #' @return An object of class 'list'. This is a named list with the following components:
 #' 
 #' \describe{
-#' \item{\code{ndi}}{An object of class 'tbl' for the GEOID, name, NDI continuous, NDI quintiles, and raw census values of specified census tracts.}
+#' \item{\code{ndi}}{An object of class 'tbl' for the GEOID, name, NDI continuous, NDI quintiles, and raw census values of specified census geographies.}
 #' \item{\code{pca}}{An object of class 'principal', returns the output of \code{\link[psych]{principal}} used to compute the NDI values.}
 #' \item{\code{missing}}{An object of class 'tbl' of the count and proportion of missingness for each census variable used to compute the NDI.}
 #' \item{\code{cronbach}}{An object of class 'character' or 'numeric' for the results of the Cronbach's alpha calculation. If only one factor is computed, a message is returned. If more than one factor is computed, the Cronbach's alpha is calculated and should check that it is >0.7 for respectable internal consistency between factors.}
@@ -70,7 +70,7 @@
 #' 
 powell_wiley <- function(geo = "tract", year, imp = FALSE, quiet = FALSE, ...) {
   
-  # Check arugments
+  # Check arguments
   match.arg(geo, choices = c("county", "tract"))
   stopifnot(is.numeric(year), year %in% 2010:2020)