+++ tests/testthat/test-messer.R
@@ -7,10 +7,10 @@ context("messer")
 test_that("messer throws error with invalid arguments", {
   
   # Unavailable geography
-  expect_error(messer(geo = "zcta", state = "DC", year = 2009, quiet = TRUE))
+  expect_error(messer(geo = "zcta", state = "DC", year = 2020, quiet = TRUE))
   
   # Unavailable year
-  expect_error(messer(state = "DC", year = 2009, quiet = TRUE))
+  expect_error(messer(state = "DC", year = 2005, quiet = TRUE))
   
   skip_if(Sys.getenv("CENSUS_API_KEY") == "")