+++ NEWS.md
@@ -1,5 +1,16 @@
 # ndi (development version)
 
+# ndi v0.1.1
+* Added `anthopolos()` function to compute the Racial Isolation Index (RI) based on based on [Anthopolos et al. (2011)](https://www.doi.org/10.1016/j.sste.2011.06.002) for specified counties/tracts 2009-2020
+* Added `bravo()` function to compute the Educational Isolation Index (EI) based on based on [Bravo et al. (2021)](https://www.doi.org/10.3390/ijerph18179384) for specified counties/tracts 2009-2020
+* Added `gini()` function to retrieve the Gini Index based on [Gini (1921)](https://www.doi.org/10.2307/2223319) for specified counties/tracts 2009-2020
+* `Matrix` and `sf` are now Depends
+* Updated vignette and README for new features
+* Fixed typos throughout documentation
+* Updated Description in DESCRIPTION
+* Updated 'package.R' with new details and section
+* Updated CITATION with new citations for the additional metrics
+
 # ndi v0.1.0
 * Fixed invalid URL and typos in package README.md